LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/html/interaction/focus/processing-model/preventScroll-textarea-expected.txt



FAIL preventScroll: true on a textarea element assert_equals: TEXTAREA: Should not have scrolled expected 0 but got 920
